Elasticsearch Tutorial
Elasticsearch is a search engine based on Lucene. It provides a distributed, multitenant-capable full-text search engine
with an HTPP web interface and schema-free JSON documents.

Is developed in Java and is released as open source under the terms of the Apache License and is the most
popular enterprise search engine followed by Apache Solr, also based on Lucene.